mv v3 config into main config pkg
