+++ pkg/bindmounts/bindmounts.go
@@ -5,7 +5,7 @@ import (
 	"io/ioutil"
 	"path/filepath"
 
-	config "github.com/craftcms/nitro/pkg/config/v3"
+	"github.com/craftcms/nitro/pkg/config"
 )
 
 func ForApp(app config.App) ([]string, error) {
@@ -14,6 +14,30 @@ func ForApp(app config.App) ([]string, error) {
 	}
 
 	return nil, fmt.Errorf("we are not done")
+}
+
+func ForSite(s config.Site, home string) ([]string, error)  {
+	// get the abs path for the site
+	path, err := s.GetAbsPath(home)
+	if err != nil {
+		return nil, err
+	}
+
+	// are there files or directories we should exclude?
+	if len(s.Excludes) > 0 {
+		var binds []string
+		for _, v := range FromDir(path, s.Excludes) {
+			_, f := filepath.Split(v)
+
+			binds = append(binds, fmt.Sprintf("%s:/app/%s:rw", v, f))
+		}
+
+		return binds, nil
+	}
+
+	// return the entire directory as the bind mount
+	return []string{fmt.Sprintf("%s:/app:rw", path)}, nil
+
 }
 
 // FromDir takes a directory path and a list of directories to exclude and returns the
