+++ pkg/config/config.go
@@ -8,8 +8,8 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/craftcms/nitro/pkg/bindmounts"
 	"github.com/craftcms/nitro/pkg/helpers"
+	"github.com/craftcms/nitro/pkg/paths"
 
 	"gopkg.in/yaml.v3"
 )
@@ -21,8 +21,8 @@ var (
 	// ErrNoConfigFile is returned when a configuration file cannot be found
 	ErrNoConfigFile = fmt.Errorf("there is no config file for the environment")
 
-	// ErrEmptyfile is returned when a config file is empty
-	ErrEmptyfile = fmt.Errorf("the config file appears to be empty")
+	// ErrEmptyFile is returned when a config file is empty
+	ErrEmptyFile = fmt.Errorf("the config file appears to be empty")
 
 	// FileName is the default name for the yaml file
 	FileName = "nitro.yaml"
@@ -54,9 +54,10 @@ type Config struct {
 	Databases  []Database  `json:"databases,omitempty" yaml:"databases,omitempty"`
 	Services   Services    `json:"services" yaml:"services"`
 	Sites      []Site      `json:"sites,omitempty" yaml:"sites,omitempty"`
-	File       string      `json:"-" yaml:"-"`
+	Apps       []App       `yaml:"apps,omitempty"`
 
-	// rw sync.RWMutex
+	File          string `json:"-" yaml:"-"`
+	HomeDirectory string `yaml:"-"`
 }
 
 // AllSitesWithHostnames takes the address, which is the nitro-proxy
@@ -155,6 +156,43 @@ func (c *Config) ListOfSitesByDirectory(home, wd string) []Site {
 	return c.Sites
 }
 
+type App struct {
+	Config     string   `yaml:"config,omitempty"`
+	Hostname   string   `yaml:"hostname,omitempty"`
+	Aliases    []string `yaml:"aliases,omitempty"`
+	Path       string   `yaml:"path,omitempty"`
+	Webroot    string   `yaml:"webroot,omitempty"`
+	PHPVersion string   `yaml:"php_version,omitempty"`
+	Dockerfile bool     `yaml:"dockerfile,omitempty"`
+	Excludes   []string `yaml:"excludes,omitempty"`
+	PHP        struct {
+		DisplayErrors             bool   `yaml:"display_errors,omitempty"`
+		MaxExecutionTime          int    `yaml:"max_execution_time,omitempty"`
+		MaxInputVars              int    `yaml:"max_input_vars,omitempty"`
+		MaxInputTime              int    `yaml:"max_input_time,omitempty"`
+		MaxFileUpload             string `yaml:"max_file_upload,omitempty"`
+		MemoryLimit               string `yaml:"memory_limit,omitempty"`
+		OpcacheEnable             bool   `yaml:"opcache_enable,omitempty"`
+		OpcacheRevalidateFreq     int    `yaml:"opcache_revalidate_freq,omitempty"`
+		OpcacheValidateTimestamps bool   `yaml:"opcache_validate_timestamps,omitempty"`
+		PostMaxSize               string `yaml:"post_max_size,omitempty"`
+		UploadMaxFileSize         string `yaml:"upload_max_file_size,omitempty"`
+	} `yaml:"php,omitempty"`
+	Extensions []string `yaml:"extensions,omitempty"`
+	Xdebug     bool     `yaml:"xdebug,omitempty"`
+	Blackfire  bool     `yaml:"blackfire,omitempty"`
+	Suspended  bool     `yaml:"suspended,omitempty"`
+	Database   struct {
+		Engine  string `yaml:"engine,omitempty"`
+		Version string `yaml:"version,omitempty"`
+	} `yaml:"database,omitempty"`
+}
+
+func (c *Config) AddApp(app App) error {
+	c.Apps = append(c.Apps, app)
+	return nil
+}
+
 // Blackfire allows users to setup their containers to use blackfire locally.
 type Blackfire struct {
 	ClientID    string `json:"client_id,omitempty" yaml:"client_id,omitempty"`
@@ -166,7 +204,7 @@ type Blackfire struct {
 // Container represents a custom container to add to nitro. Containers can be
 // publicly hosted on Docker Hub.
 type Container struct {
-	// Name is a uniq name, with no spaces or special characters and is used to generate the hostname
+	// Name is a unique name, with no spaces or special characters and is used to generate the hostname
 	Name string `json:"name" yaml:"name"`
 
 	// Image the is canonical name of the image to use for the container `docker.elastic.co/elasticsearch/elasticsearch`
@@ -292,31 +330,6 @@ func (s *Site) GetAbsPath(home string) (string, error) {
 	return cleanPath(home, s.Path)
 }
 
-// GetBindMounts takes the users home directory and will return a
-// list of bind mounts that checks the excludes on the site.
-func (s *Site) GetBindMounts(home string) ([]string, error) {
-	// get the abs path for the site
-	path, err := s.GetAbsPath(home)
-	if err != nil {
-		return nil, err
-	}
-
-	// are there files or directories we should exclude?
-	if len(s.Excludes) > 0 {
-		var binds []string
-		for _, v := range bindmounts.FromDir(path, s.Excludes) {
-			_, f := filepath.Split(v)
-
-			binds = append(binds, fmt.Sprintf("%s:/app/%s:rw", v, f))
-		}
-
-		return binds, nil
-	}
-
-	// return the entire directory as the bind mount
-	return []string{fmt.Sprintf("%s:/app:rw", path)}, nil
-}
-
 // GetAbsContainerPath gets the directory for a site’s
 // container path.
 func (s *Site) GetAbsContainerPath(home string) (string, error) {
@@ -324,7 +337,7 @@ func (s *Site) GetAbsContainerPath(home string) (string, error) {
 }
 
 // GetContainerPath is responsible for looking at the
-// site’s web root and determing the correct path in the
+// site’s web root and determining the correct path in the
 // container. This is used for the craft and queue
 // commands to identify the location of the "craft"
 // executable.
@@ -538,7 +551,8 @@ func Load(home string) (*Config, error) {
 
 	// create the config
 	c := &Config{
-		File: file,
+		File:          file,
+		HomeDirectory: home,
 	}
 
 	// read the file
@@ -551,6 +565,106 @@ func Load(home string) (*Config, error) {
 	if err := yaml.Unmarshal(data, &c); err != nil {
 		return nil, err
 	}
+
+	// load each of the apps
+	for i, global := range c.Apps {
+		// if there is a config file, load it
+		if global.Config != "" {
+			// load the file
+			p, err := paths.Clean(c.HomeDirectory, global.Config)
+			if err != nil {
+				return nil, err
+			}
+
+			// read the local app config file
+			local, err := unmarshalAppConfigFrom(p)
+			if err != nil {
+				return nil, err
+			}
+
+			// parse the values but global values override the global config if present
+
+			// check the hostname
+			if hostname, err := parseStringValue(global, local, "hostname"); err == nil {
+				c.Apps[i].Hostname = hostname
+			} else if err != nil {
+				return c, err
+			}
+
+			// check the aliases
+			if global.Aliases != nil {
+				c.Apps[i].Aliases = global.Aliases
+			} else if local.Aliases != nil {
+				c.Apps[i].Aliases = local.Aliases
+			}
+
+			// check the webroot
+			if webroot, err := parseStringValue(global, local, "webroot"); err == nil {
+				c.Apps[i].Webroot = webroot
+			} else if err != nil {
+				return c, err
+			}
+
+			// check the php_version
+			if phpVersion, err := parseStringValue(global, local, "php_version"); err == nil {
+				c.Apps[i].PHPVersion = phpVersion
+			} else if err != nil {
+				return c, err
+			}
+
+			// check the dockerfile
+			if global.Dockerfile != local.Dockerfile {
+				c.Apps[i].Dockerfile = global.Dockerfile
+			} else {
+				c.Apps[i].Dockerfile = local.Dockerfile
+			}
+
+			// TODO(jasonmccallister) check the php settings
+
+			// check the php extensions
+			if global.Extensions != nil {
+				c.Apps[i].Extensions = global.Extensions
+			} else if local.Extensions != nil {
+				c.Apps[i].Extensions = local.Extensions
+			}
+
+			// check xdebug
+			if global.Xdebug != local.Xdebug {
+				c.Apps[i].Xdebug = global.Xdebug
+			} else {
+				c.Apps[i].Xdebug = local.Xdebug
+			}
+
+			// check blackfire
+			if global.Blackfire != local.Blackfire {
+				c.Apps[i].Blackfire = global.Blackfire
+			} else {
+				c.Apps[i].Blackfire = local.Blackfire
+			}
+
+			// check suspend
+			if global.Suspended != local.Suspended {
+				c.Apps[i].Suspended = global.Suspended
+			} else {
+				c.Apps[i].Suspended = local.Suspended
+			}
+
+			// check the database engine
+			if global.Database.Engine != "" {
+				c.Apps[i].Database.Engine = global.Database.Engine
+			} else if local.Database.Engine != "" {
+				c.Apps[i].Database.Engine = local.Database.Engine
+			}
+
+			// check the database version
+			if global.Database.Version != "" {
+				c.Apps[i].Database.Version = global.Database.Version
+			} else if local.Database.Version != "" {
+				c.Apps[i].Database.Version = local.Database.Version
+			}
+		}
+	}
+
 
 	// return the config
 	return c, nil
@@ -567,7 +681,7 @@ func IsEmpty(home string) (string, error) {
 
 	// check if the file is empty
 	if stat.Size() == 0 {
-		return "", ErrEmptyfile
+		return "", ErrEmptyFile
 	}
 
 	return file, nil
@@ -865,4 +979,56 @@ func cleanPath(home, path string) (string, error) {
 	}
 
 	return filepath.Clean(abs), nil
+}
+
+func parseStringValue(global, local App, key string) (string, error) {
+	switch key {
+	case "hostname":
+		if global.Hostname != "" {
+			return global.Hostname, nil
+		}
+
+		if local.Hostname != "" {
+			return local.Hostname, nil
+		}
+
+		return "", fmt.Errorf("hostname must be defined for an app")
+	case "php_version":
+		if global.PHPVersion != "" {
+			return global.PHPVersion, nil
+		}
+
+		if local.PHPVersion != "" {
+			return local.PHPVersion, nil
+		}
+
+		return "", fmt.Errorf("php_version must be defined for an app")
+	case "webroot":
+		if global.Webroot != "" {
+			return global.Webroot, nil
+		}
+
+		if local.Webroot != "" {
+			return local.Webroot, nil
+		}
+
+		return "", fmt.Errorf("webroot must be defined for an app")
+	}
+
+	return "", fmt.Errorf("unkown key %q provided", key)
+}
+
+func unmarshalAppConfigFrom(path string) (App, error) {
+	var app App
+	f, err := ioutil.ReadFile(path)
+	if err != nil {
+		return app, fmt.Errorf("unable to read file at %v", path)
+	}
+
+	err = yaml.Unmarshal(f, &app)
+	if err != nil {
+		return app, fmt.Errorf("unable to marshal file, %v", err)
+	}
+
+	return app, nil
 }
