--- pkg/config/testdata/path-with-config/nitro.yaml
@@ -1,1 +1,0 @@
-hostname: custom-hostname-from-file.nitro
