+++ pkg/config/testdata/team-site-with-config/nitro.yaml
@@ -1,4 +1,3 @@
 hostname: team-site-name-from-config.nitro
 php_version: "7.4"
 webroot: public
-
