--- pkg/config/v3/testdata/.nitro/nitro.yaml
@@ -1,11 +1,0 @@
-apps:
-  - hostname: mysite.nitro
-    path: ~/mysite
-    php_version: "7.4"
-    webroot: web
-  - config: ~/team-site-with-config/nitro.yaml
-    aliases:
-      - my-local-app.test
-    php_version: "8.0"
-    extensions:
-      - grpc
