--- pkg/config/v3/testdata/mysite/.gitignore
@@ -1,2 +1,0 @@
-.gitignore
-!.gitignore
