+++ pkg/customcontainer/customcontainer.go
@@ -8,8 +8,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/craftcms/nitro/command/apply/internal/match"
-	config "github.com/craftcms/nitro/pkg/config/v3"
+	"github.com/craftcms/nitro/pkg/config"
 	"github.com/craftcms/nitro/pkg/containerlabels"
 	"github.com/craftcms/nitro/pkg/pathexists"
 	"github.com/docker/docker/api/types"
@@ -57,23 +56,25 @@ func StartOrCreate(ctx context.Context, docker client.CommonAPIClient, home, net
 		return "", err
 	}
 
+	return "", fmt.Errorf("not yet implemented", details)
+
 	// if the container is out of date
-	if err := match.Container(home, c, details); err != nil {
-		fmt.Println(err)
-		fmt.Print("- updating… ")
+	// if err := match.Container(home, c, details); err != nil {
+	// 	fmt.Println(err)
+	// 	fmt.Print("- updating… ")
 
-		// stop container
-		if err := docker.ContainerStop(ctx, container.ID, nil); err != nil {
-			return "", err
-		}
+	// 	// stop container
+	// 	if err := docker.ContainerStop(ctx, container.ID, nil); err != nil {
+	// 		return "", err
+	// 	}
 
-		// remove container
-		if err := docker.ContainerRemove(ctx, container.ID, types.ContainerRemoveOptions{}); err != nil {
-			return "", err
-		}
+	// 	// remove container
+	// 	if err := docker.ContainerRemove(ctx, container.ID, types.ContainerRemoveOptions{}); err != nil {
+	// 		return "", err
+	// 	}
 
-		return create(ctx, docker, home, networkID, c)
-	}
+	// 	return create(ctx, docker, home, networkID, c)
+	// }
 
 	return container.ID, nil
 }
